+++ script/character/player.ts
@@ -417,7 +417,7 @@ const defaultPlayer = {
   level: 1,
   xp: 0,
   starting_aspect: "strength",
-  class: classManager.get("warrior"),
+  class: classManager.get("mage"),
 };
 
 let player = new Player({