+++ script/compiled/character/class_perk.js
@@ -3,6 +3,7 @@ class ClassPerk {
     id;
     type;
     class;
+    exclusive;
     unlock;
     modifiers;
     commands;
@@ -10,25 +11,34 @@ class ClassPerk {
         this.id = base.id;
         this.type = base.type;
         this.class = base.class;
+        this.exclusive = base.exclusive;
         this.unlock = base.unlock;
         this.modifiers = base.modifiers;
         this.commands = base.commands;
     }
     available() {
+        if (DEVTOOLS.IGNORE_REQUIREMENTS)
+            return true;
+        if (player.hasClassPerk(this.id))
+            return false;
+        if (this.exclusive) {
+            for (let perkId of this.exclusive) {
+                if (player.hasClassPerk(perkId))
+                    return false;
+            }
+        }
         if (this.unlock.stats) {
             const stats = this.unlock.stats;
             const pStats = player.getStats();
-            Object.keys(stats).forEach((stat) => {
+            for (let stat in stats) {
                 if (pStats[stat] < stats[stat])
                     return false;
-            });
+            }
         }
         if (this.unlock.gold) {
             if (player.gold < this.unlock.gold)
                 return false;
         }
-        if (player.class.perks.find((perk) => perk.id === this.id))
-            return false;
         console.log("lmao");
         return true;
     }
@@ -38,7 +48,7 @@ class ClassPerk {
         hideHover();
         closeConfirmationWindow();
         if (!options?.confirmed) {
-            return confirmationWindow(`<f>1.5rem<f>${game.getLocalizedString("confirm_perk_unlock").replace("{id}", this.id)}`, () => this.assign({ confirmed: true }));
+            return confirmationWindow(`<f>1.5rem<f>${game.getLocalizedString("confirm_perk_unlock").replace("{id}", game.getLocalizedString(this.id))}`, () => this.assign({ confirmed: true }));
         }
         player.removeGold(this.unlock.gold);
         player.class.perks.push({ ...this });
@@ -60,6 +70,14 @@ class ClassPerk {
         tooltip += `<c>silver<c>"${game.getLocalizedString(this.id + "_desc")}"<c>white<c>\n`;
         if (player.hasClassPerk(this.id)) {
             tooltip += `<c>lime<c>${game.getLocalizedString("class_perk_owned")}<c>white<c>\n`;
+        }
+        if (this.exclusive) {
+            tooltip += `<c>white<c>${game.getLocalizedString("class_perk_exclusive")}:`;
+            this.exclusive.forEach((perkId) => {
+                const col = player.hasClassPerk(perkId) ? "red" : "goldenrod";
+                tooltip += ` <c>${col}<c>${game.getLocalizedString(perkId)}<c>white<c>`;
+            });
+            tooltip += "\n";
         }
         // Perk unlock
         if (this.unlock.stats) {
@@ -131,7 +149,93 @@ const classPerks = {
                         strP: 5,
                     },
                 }),
+                ,
             ],
+        },
+        {
+            level: 5,
+            perks: [
+                new ClassPerk({
+                    id: "warrior_toughness",
+                    type: "classPerk",
+                    class: "warrior",
+                    unlock: {
+                        stats: {
+                            vit: 20,
+                        },
+                        gold: 2000,
+                    },
+                    modifiers: {
+                        hpMaxFromStrV: 0.5,
+                        vitP: 5,
+                    },
+                }),
+                {
+                    exclusive: true,
+                    perks: [
+                        new ClassPerk({
+                            id: "warrior_battle_arts",
+                            type: "classPerk",
+                            class: "warrior",
+                            exclusive: ["warrior_berserking"],
+                            unlock: {
+                                stats: {
+                                    int: 15,
+                                    vit: 15,
+                                },
+                                gold: 1500,
+                            },
+                            modifiers: {
+                                spellPowerP: 10,
+                                healPowerP: 7.5,
+                                intP: 5,
+                            },
+                        }),
+                        new ClassPerk({
+                            id: "warrior_berserking",
+                            type: "classPerk",
+                            class: "warrior",
+                            exclusive: ["warrior_battle_arts"],
+                            unlock: {
+                                stats: {
+                                    str: 20,
+                                    vit: 15,
+                                },
+                                gold: 1500,
+                            },
+                            modifiers: {
+                                meleeDamageP: 12.5,
+                                strP: 10,
+                                strV: 5,
+                                vitV: 3,
+                                intV: -10,
+                                spellPowerP: -10,
+                                healPowerP: -7.5,
+                            },
+                        }),
+                    ],
+                },
+            ],
+        },
+        {
+            level: 10,
+            perks: [],
+        },
+        {
+            level: 15,
+            perks: [],
+        },
+        {
+            level: 20,
+            perks: [],
+        },
+        {
+            level: 25,
+            perks: [],
+        },
+        {
+            level: 30,
+            perks: [],
         },
     ],
     rogue: [
@@ -176,6 +280,88 @@ const classPerks = {
                 }),
             ],
         },
+        {
+            level: 5,
+            perks: [
+                new ClassPerk({
+                    id: "rogue_critical_strike",
+                    type: "classPerk",
+                    class: "rogue",
+                    unlock: {
+                        stats: {
+                            agi: 20,
+                        },
+                        gold: 2000,
+                    },
+                    modifiers: {
+                        critPowerV: 15,
+                        critRateV: 5,
+                        dodgeV: 2,
+                        poisonResistanceV: 10,
+                    },
+                }),
+                {
+                    exclusive: true,
+                    perks: [
+                        new ClassPerk({
+                            id: "rogue_bob_and_weave",
+                            type: "classPerk",
+                            class: "rogue",
+                            unlock: {
+                                stats: {
+                                    agi: 20,
+                                },
+                                gold: 1500,
+                            },
+                            modifiers: {
+                                dodgeV: 10,
+                                meleeDamageP: 10,
+                                critPowerV: 10,
+                                critRateV: 5,
+                                hpMaxP: -15,
+                            },
+                        }),
+                        new ClassPerk({
+                            id: "rogue_parry_and_riposte",
+                            type: "classPerk",
+                            class: "rogue",
+                            unlock: {
+                                stats: {
+                                    agi: 20,
+                                },
+                                gold: 1500,
+                            },
+                            modifiers: {
+                                meleeDamageP: 5,
+                                physicalDefenceV: 5,
+                                critPowerV: 5,
+                                critRateV: 2.5,
+                            },
+                        }),
+                    ],
+                },
+            ],
+        },
+        {
+            level: 10,
+            perks: [],
+        },
+        {
+            level: 15,
+            perks: [],
+        },
+        {
+            level: 20,
+            perks: [],
+        },
+        {
+            level: 25,
+            perks: [],
+        },
+        {
+            level: 30,
+            perks: [],
+        },
     ],
     mage: [
         {
@@ -218,6 +404,88 @@ const classPerks = {
                     },
                 }),
             ],
+        },
+        {
+            level: 5,
+            perks: [
+                new ClassPerk({
+                    id: "mage_mana_reserves",
+                    type: "classPerk",
+                    class: "mage",
+                    unlock: {
+                        stats: {
+                            int: 20,
+                        },
+                        gold: 2000,
+                    },
+                    modifiers: {
+                        mpMaxFromIntV: 1,
+                        mpMaxP: 10,
+                        mpRegenP: 10,
+                    },
+                }),
+                {
+                    exclusive: true,
+                    perks: [
+                        new ClassPerk({
+                            id: "mage_blast_them",
+                            type: "classPerk",
+                            class: "mage",
+                            exclusive: ["mage_conserve_mana"],
+                            unlock: {
+                                stats: {
+                                    int: 20,
+                                },
+                                gold: 1500,
+                            },
+                            modifiers: {
+                                spellPowerP: 25,
+                                healPowerP: 10,
+                                magicalDamageP: 5,
+                                elementalDamageP: 5,
+                                all_mpCostP: 30,
+                                all_cooldownV: 1,
+                            },
+                        }),
+                        new ClassPerk({
+                            id: "mage_conserve_mana",
+                            type: "classPerk",
+                            class: "mage",
+                            exclusive: ["mage_blast_them"],
+                            unlock: {
+                                stats: {
+                                    int: 20,
+                                },
+                                gold: 1500,
+                            },
+                            modifiers: {
+                                all_mpCostP: -15,
+                                all_cooldownP: -10,
+                            },
+                        }),
+                    ],
+                },
+            ],
+        },
+        {
+            level: 10,
+            perks: [],
+        },
+        {
+            level: 15,
+            perks: [],
+        },
+        {
+            level: 20,
+            perks: [],
+        },
+        {
+            level: 25,
+            perks: [],
+        },
+        {
+            level: 30,
+            perks: [],
         },
     ],
     paladin: [
@@ -262,6 +530,93 @@ const classPerks = {
                     },
                 }),
             ],
+        },
+        {
+            level: 5,
+            perks: [
+                new ClassPerk({
+                    id: "paladin_blessing",
+                    type: "classPerk",
+                    class: "paladin",
+                    unlock: {
+                        stats: {
+                            spi: 15,
+                            vit: 15,
+                        },
+                        gold: 2000,
+                    },
+                    modifiers: {
+                        healPowerP: 15,
+                        vitP: 7.5,
+                    },
+                }),
+                {
+                    exclusive: true,
+                    perks: [
+                        new ClassPerk({
+                            id: "paladin_divine_dedication",
+                            type: "classPerk",
+                            class: "paladin",
+                            exclusive: ["paladin_sword_and_shield"],
+                            unlock: {
+                                stats: {
+                                    spi: 20,
+                                },
+                                gold: 1500,
+                            },
+                            modifiers: {
+                                spellPowerP: 12.5,
+                                mpMaxP: 5,
+                                healPowerP: 2.5,
+                                spiV: 2,
+                                intV: 2,
+                                hpMaxP: -5,
+                            },
+                        }),
+                        new ClassPerk({
+                            id: "paladin_sword_and_shield",
+                            type: "classPerk",
+                            class: "paladin",
+                            exclusive: ["paladin_divine_dedication"],
+                            unlock: {
+                                stats: {
+                                    str: 15,
+                                    vit: 20,
+                                },
+                                gold: 1500,
+                            },
+                            modifiers: {
+                                meleeDamageP: 10,
+                                hpMaxP: 5,
+                                strV: 2,
+                                vitV: 2,
+                                spellPowerP: -5,
+                                mpMaxP: -5,
+                            },
+                        }),
+                    ],
+                },
+            ],
+        },
+        {
+            level: 10,
+            perks: [],
+        },
+        {
+            level: 15,
+            perks: [],
+        },
+        {
+            level: 20,
+            perks: [],
+        },
+        {
+            level: 25,
+            perks: [],
+        },
+        {
+            level: 30,
+            perks: [],
         },
     ],
 };