+++ script/compiled/lobby/char_view.js
@@ -74,10 +74,10 @@ function createCharView() {
         const breakdown = getModifierBreakdown(key, player);
         if (breakdown) {
             let breakdownText = `${game.getLocalizedString("breakdown")}:\n`;
-            Object.entries(breakdown).forEach(([key, value]) => {
-                if (value === 1)
+            Object.entries(breakdown).forEach(([_key, _value]) => {
+                if (_value === 1)
                     return;
-                breakdownText += effectSyntax(key, (value - 1) * 100);
+                breakdownText += effectSyntax(_key, (_value - 1) * 100, getProperties(key));
             });
             tooltip(modifier, breakdownText);
         }