+++ script/dev/dev.ts
@@ -12,14 +12,15 @@ const DEVTOOLS: DevTools = {
   NO_CD: false,
   FREE_CAST: false,
   GOD: false,
+  IGNORE_REQUIREMENTS: false,
 };
 
 if (localStorage.getItem("devtools") === "true") DEVTOOLS.ENABLED = true;
 
-const gameVersion = (0.13).toFixed(2);
-const gameVersionText = (ver: string) => {
+const gameVersion = (0.14).toFixed(2);
+const gameVersionText = (ver: string): string => {
   if (parseFloat(ver) < 0.1) return "Old";
-  return `${parseInt(ver)}.${(parseFloat(ver) * 10) % 10}`;
+  return `${parseInt(ver)}.${((parseFloat(ver) * 10) % 10).toFixed(1)}`;
 };
 
 // Skip main menu