+++ script/perk/skill.ts
@@ -73,6 +73,7 @@ class Skill {
 
   available(): boolean {
     let available = true;
+    if (DEVTOOLS.IGNORE_REQUIREMENTS) return true;
     if (player.skill_points <= 0) return false;
     if (this.requirements) {
       this.requirements.forEach((req: any) => {